Split commands-generator-api
Since android plugin cannot be applied together with withJava, before applying the android plugin we need to remove all java code from the applied kmp module. This forces the creation of a new java module (not kmp) which will contain all the jvm specific code for commands-generator-api. This is needed even though there's no java code there, because we want the dagger generated java code to be included.

This also forces the creation of another module (commands-generator-api-common) which contains all the code that's required both by the code in the new java module, but by non-jvm variants